Refactor: Integrate service layer with controller and add transactional management
