+++ src/main/java/com/portfolio/livros/controller/LivroController.java
@@ -1,9 +1,8 @@
 package com.portfolio.livros.controller;
 
-import com.portfolio.livros.model.livro.DadosCadastraLivro;
-import com.portfolio.livros.model.livro.DadosEditarLivro;
-import com.portfolio.livros.model.livro.Livro;
-import com.portfolio.livros.model.livro.LivroRepository;
+import com.portfolio.livros.model.DadosCadastraLivro;
+import com.portfolio.livros.model.DadosEditarLivro;
+import com.portfolio.livros.service.LivroService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -15,12 +14,12 @@ import org.springframework.web.bind.annotation.*;
 public class LivroController {
 
     @Autowired
-    private LivroRepository repository;
+    private LivroService livroService;
 
     @GetMapping("formulario")
     public String carregaFormulario(Long id, Model model) {
         if (id != null) {
-            var livro = repository.getReferenceById(id);
+            var livro = livroService.findById(id);
             model.addAttribute("livro", livro);
         }
 
@@ -29,20 +28,20 @@ public class LivroController {
 
     @GetMapping
     public String carregaLivros(Model model) {
-        model.addAttribute("lista", repository.findAll());
+        model.addAttribute("lista", livroService.findAll());
         return "livros/lista";
     }
+
     @PostMapping
     public String cadastraLivro(DadosCadastraLivro dadosCadastraLivro) {
-        var livro = new Livro(dadosCadastraLivro);
-        repository.save(livro);
+        livroService.save(dadosCadastraLivro);
         return "redirect:/livros";
     }
 
     @DeleteMapping
     @Transactional
     public String deleteLivro(Long id){
-        repository.deleteById(id);
+        livroService.deleteById(id);
         System.out.println("excluido");
 
         return "redirect:/livros";
@@ -51,11 +50,7 @@ public class LivroController {
     @PutMapping
     @Transactional
     public String editarLivro(DadosEditarLivro dadosEditarLivro){
-        var livro = repository.getReferenceById(dadosEditarLivro.id());
-        livro.atualizarLivro(dadosEditarLivro);
-
+        livroService.update(dadosEditarLivro);
         return "redirect:/livros";
-
     }
-
 }
