+++ src/main/java/com/portfolio/livros/model/DadosCadastraLivro.java
@@ -1,4 +1,4 @@
-package com.portfolio.livros.model.livro;
+package com.portfolio.livros.model;
 
 public record DadosCadastraLivro(String titulo, String autor) {
 }
