+++ src/main/java/com/portfolio/livros/model/DadosEditarLivro.java
@@ -1,4 +1,4 @@
-package com.portfolio.livros.model.livro;
+package com.portfolio.livros.model;
 
 public record DadosEditarLivro(Long id, String titulo, String autor) {
 }
