+++ src/main/resources/application.properties
@@ -1,5 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost/meuslivros?createDatabaseIfNotExist=true
+spring.jpa.hibernate.ddl-auto=update
 spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.password=hard234
 
 spring.mvc.hiddenmethod.filter.enabled=true
